refactor: introduce node module in nitro extension
