+++ extensions/inference-nitro-extension/src/@types/global.d.ts
@@ -1,4 +1,4 @@
-declare const MODULE: string;
+declare const NODE: string;
 declare const INFERENCE_URL: string;
 declare const TROUBLESHOOTING_URL: string;
